+++ libs/ui/src/lib/static-autocomplete/group/group.component.ts
@@ -15,9 +15,9 @@ import { BehaviorSubject, combineLatest, Observable, Subscription, defer } from 
 import { map, startWith, shareReplay, pairwise } from "rxjs/operators";
 import { Scope, StaticGroup, staticGroups, staticModel } from '@blockframes/utils/static-model';
 import { boolean } from '@blockframes/utils/decorators/decorators';
+import { GroupMode } from "@blockframes/utils/pipes/get-mode.pipe";
 
 
-type GroupMode = 'indeterminate' | 'checked' | 'unchecked';
 
 function filter([groups, text]: [StaticGroup[], string], scope: Scope) {
   if (!text) return groups;
@@ -38,17 +38,7 @@ function filter([groups, text]: [StaticGroup[], string], scope: Scope) {
   return result;
 }
 
-function getMode(group: StaticGroup, value: string[]): GroupMode {
-  const items = group.items.filter(item => value.includes(item));
-  if (!items.length) return 'unchecked';
-  if (items.length === group.items.length) return 'checked';
-  return 'indeterminate';
-}
-function getRootMode(groups: StaticGroup[], value: string[]): GroupMode {
-  if (!value.length) return 'unchecked';
-  if (groups.every(group => getMode(group, value) === 'checked')) return 'checked';
-  return 'indeterminate';
-}
+
 
 function getItems(groups: StaticGroup[]): string[] {
   return groups.reduce((items, group) => items.concat(group.items), []);
@@ -198,13 +188,6 @@ export class StaticGroupComponent implements ControlValueAccessor, OnInit, OnDes
 }
 
 
-@Pipe({ name: 'getMode' })
-export class GetModePipe implements PipeTransform {
-  transform(value: string[], group: StaticGroup | StaticGroup[]) {
-    if (Array.isArray(group)) return getRootMode(group, value);
-    return getMode(group, value);
-  }
-}
 
 @Pipe({ name: 'triggerDisplay' })
 export class TriggerDisplayValue implements PipeTransform {
