+++ libs/ui/src/lib/static-autocomplete/group/group.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { StaticGroupComponent, GetModePipe, TriggerDisplayValue } from './group.component';
+import { StaticGroupComponent, TriggerDisplayValue } from './group.component';
 import { ReactiveFormsModule } from "@angular/forms";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatSelectModule } from "@angular/material/select";
@@ -11,7 +11,7 @@ import { MatIconModule } from "@angular/material/icon";
 import { MatDividerModule } from "@angular/material/divider";
 import { MatRippleModule } from "@angular/material/core";
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { ToLabelModule } from '@blockframes/utils/pipes';
+import { GetModePipeModule, ToLabelModule } from '@blockframes/utils/pipes';
 
 @NgModule({
   imports: [
@@ -22,13 +22,14 @@ import { ToLabelModule } from '@blockframes/utils/pipes';
     MatButtonModule,
     MatIconModule,
     MatCheckboxModule,
+    GetModePipeModule,
     MatInputModule,
     MatDividerModule,
     MatRippleModule,
     MatTooltipModule,
     ToLabelModule,
   ],
-  declarations: [StaticGroupComponent, GetModePipe, TriggerDisplayValue],
+  declarations: [StaticGroupComponent, TriggerDisplayValue],
   exports: [StaticGroupComponent]
 })
 export class StaticGroupModule { }
